Add delete command/api endpoint
This was removed in the state change API (because the rest endpoint changed
slightly). Here we re-add it.

I've also changed the delete mode to be default destructive, i.e. you can
delete a resource that is in use. I am often supremely annoyed when I get these
warnings, which is why I did it. I don't mind adding a --use-caution flag if we
want to keep the original behavior, but I think for ease of use, it would be
nice to have the default be destructive.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>