Implement GET /1.0 API
The most interesting bit here is probably the filesystem detection bit. I'm not
entirely sure of the best way of doing the int => string conversion; as near as
my google-fu could tell, there isn't a library for doing this sort of thing, so
I just put some interesting filesystems in.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>