Actually return a 404 for a missing container
NewContainer always returns a new container (unless some internal LXC error has
happend), even if it is not defined yet. In some cases, we only want to allow
operations on containers that exist, so we should check and be sure the
container is defined before we try to perform operations on it.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>