Move container name pasring out of NewClient
There was a slight bug here: when doing `lxc list foo`, lxc would always list
the default remote, since the previous parsing assumed that "foo" was the
container name. Instead, we should be smarter about when we're parsing what
kind of name, and not do any of it inside NewClient().

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>