store certs in memory by hostname
Instead of retroactively chopping off the extension, we should just store
certs initially by hostname instead of by file name. When we got a new cert,
we would compare it against a cert without the extension, and thus end up
adding duplicate certs in memory (but not on disk, since we appended the
filename there). Instead, we should just store and look up everything by
hostname.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>