Add wait-for-websocket parameter to exec
The problem here was a race: if we didn't wait for the user to connect to the
websocket before running the command, the command could finish before the user
even connected, and we wouldn't send it any input or receive any output.
However, sometimes it would be nice to run a command without having to connect
to the websocket interface, so we support that as well.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>