Add support for internationalization
On small caveat here is that xgettext doesn't actually support golang at the
moment, so we are hacking it by using C++ as the language. Unfortunately, this
means we have to mangle our multiline strings into single line strings, which
is rather ugly.

Additionally, we've decided for now not to i18n lxd itself (in favor of just
the client). There are several problems here: 1. how do we determine what
locale to render messages in? (Accept-Encoding is probably the answer here),
and 2. how do we do the translation with gettext in go in multithreaded land?
I'm not sure on #2, so we've punted for now.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>