Fix race in tomb invocation
Tombs die afteir all of their managed goroutines die. In this case, if the TCP
listener fails to bind and die, its goroutine could return and cause the tomb
to die. If that all happens before the next line (the unix socket listener) is
started in the tomb, then there is a panic. Instead, we start the socket
listeners inside of a goroutine themselves, so that the tomb is guaranteed to
be alive while both of the goroutines are starting.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>