Replace make_lxc_tarball.sh with scripts/lxd-images
This new script written in python3 will let you import any image from
https://images.linuxcontainers.org, repack it into lxd's format and then
import it into the local LXD's image store.

At this point, the script is designed to run as a user with access to
the lxd socket (typically any member of the lxd group).

Aliases can be setup by passing --alias one or multiple times.

An example is:
 scripts/lxd-images import lxc ubuntu trusty amd64 \
    --alias ubuntu --alias ubuntu/trusty --alias ubuntu/trusty/amd64

With that done, you can then start containers with any of:
 lxc launch ubuntu
 lxc launch ubuntu/trusty
 lxc launch ubuntu/trusty/amd64

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>