build fuidmap
make fuidmap/main.go package main, and move fuidmap/idmap.go into
package shared as uidmapset.go.

We should clean this stuff up so that there is a uidmapEntry and a
uidmapSet, where uidmapSet is a set of uidmapEntry's and the
uidmaapSet.ShiftUid() and uidmapSet.ShiftTree() are the helpers
which do the work.  Hopefully I'll get to that soon, but this
at least gives us a fuidmap binary again.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>