images: set properties into the db
All this does is take prop=value elements from the lxc import tarball
command and get them into the db, and delete them when the image is
deleted.

Searching at 'image list' based on properties, and editing them with
"lxc image edit", are still todo, and should at this point be quite
easy.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>