Run each websocket's Do() exactly once
Before we were running a copy of the websocket's Do() method for each websocket
that was connected. This didn't manifest any problems since each websocket
waited for multiple connections and only ran one copy, but it did leak a lot of
goroutines.

The documentation was also incorrect (in that the Do() method is actually run
before there are any connections), so clarify this for users.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>