Implement shared.dbQueryScan
This retries when the database is locked.

Unlike rows.Scan(), it takes a slice of interfaces which contain the
variable types we want, not pointers to variables we want.  The
answers are returned as a return value which is an array of arrays
of interfaces which container pointers to the values we want.
It only handles strings and ints right now - because that's all we
use.

Switch over almost all rows.Scan()s.  The only ones not converted
are only run at lxd startup.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>