bump WebsocketMirror done channel buffer size
Typical use of WebsocketMirror in our code doesn't care about when the mirror
is done. However, since we send done from both the read and write channel, if
the buffer size is 1 and nobody ever reads from the channel, the second done
signal blocks waiting for space in the channel buffer.

If we bump this to two, everything is successfully written/buffered and the
channel goes out of scope and is freed if nobody on the other end cares about
it, which is the common case.

This prevents goroutines from leaking as in:

http://people.canonical.com/~rharper/lxd/issues/522/lxd_backtrace.log

(grep for "minutes")

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>