Fix a race in hwaddr setting
Two competing newLxdContainers (for instance) can race setting a
hwaddr for a device.  This leads to a db error because the containerid
and config key must together be unique.

also add a new testcase, which is not automatically run.  We can
reconsider that and rename it, but for now it's intended to keep
growing into the most painful testcaes for lxd.

It launches 100 containers (5 waves of 20) while doing other things
in parallel.  In particular, adding 'exec' would cause this to fail
because the 'start' part of 'launch' is still unreliable.  That is
obviously next on my list.

It is currently a separate script in part because it uses bash
arrays to track the pids.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>