render both local and expanded config in /1.0/containers/<name>
Since one of the intended uses for the output of GET is to update it so that
it can be PUT back to set container settings, we should render only the local
settings. We add the new keys expanded_* so that people can still see the
expanded config if they want.

Closes #568

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>