rearrange when the GC is called
In particular, call it before the request is rendered. I think the fdleak test
was failing in travis because the request was rendered and the test finishes
running before the lxd process gets CPU time and does the garbage collection.

Of course, I might be crazy, so let's run travis a few times before we merge
this to see that it doesn't still race.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>