skip fdleak test on broken lxc
The lxc_log_fd is per-thread in older lxc (including in the stable ppa
and vivid).  This means that each separate go thread will open a new log
fd, though a re-used thread will close an old fd to open a new one.

This is fixed in lxc by the per-thread logging patch, so just disable
the fdleak test for now if we are running known-broken lxc.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>