implement lxc info <resource> --show-log
In a lot of cases it's useful look at the container's log; let's export that
via the api as well.

We add this as a query parameter instead of doing it all the time so that we
avoid opening the file unless people really want to see the log info.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>