pre-extract an image before copy/migration
If the target LXD server happens to have the base image for container that is
being migrated to it, we should extract the base image before we try to sync
the filesystem. The idea here is that a lot of the files will be the same, so
we can save some network bandwidth and rsyncing time if we keep around the
rootfs.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>