Added data migration for ON DELETE CASCADE
The previous  branhc introducing cascading functionlaity forgot the
database schema migration code, so this commit introduces such a
migration.

The logic is: for each table with future cascade behavior: create a
temporary table with all the "real" table's data. Delete the real table.
Recreate the real table with the modified schema (adding the ON DELETE
CASCADE on foreign keys), then repopulate the table with the contents of
the temporary table.

Signed-off-by: Christopher Glass <christopher.glass@canonical.com>