Use the right name for certificates
Instead of the request.TLS.ServerName, we should use the actual client's
hostname. Further, tracking (and rendering) hostnames as paired with
certificates doesn't really make sense, because multiple users wouldn't be able
to have different certificates on the same client host. So, let's not track
certs by hostname and just use the fingerprint instead.

This broke the implementation of the API, but since it was wrong (the spec says
nothing about it being a map from host:fingerprint, just that it's a list of
fingerprints), I think that's ok.

Closes #623

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>