Refactor dbImageGet to take a db pointer
Instead of passing the whole daemon to it, simply pass the db pointer.
This is a first step in making the database access layer decoupled from
the daemon and http interface, and has the desirable side effect of
making the code more testable.

The function also received a significant refactoring, dropping some dead
code.

Signed-off-by: Chris Glass <tribaal@gmail.com>