More data layer fixes
Mostly focused on decoupling the data layer from the rest by passing in
references to the sql.DB instead of the daemon.

Another instance of removing a SQL statement (preemtively counting is
not necessary, just let it fail when doing the actual work).

Refactored the db_test.go file to be a little more DRY.

Signed-off-by: Chris Glass <tribaal@gmail.com>