Detect non-running Unix socket servers
Instead of showing a cryptic error message, hint that the server might
not be running when getting a dial error on unix sockets.

Note: I would be better to switch on the err type in main instead of
exiting(1) in client.go, but the error type gets somehow mangeled along
the way, upcasted to the general "error" type. This branch gets the job
done, even if it's a bit less elegant.

Signed-off-by: Chris Glass <tribaal@gmail.com>