lxd-client should not depend on liblxc
(or sqlite3)

The Juju folks are interested in using our go client library, but as we had
things factored they needed to link against liblxc, which they didn't want to
do. This seems fairly reasonable, as liblxc brings in a lot of other deps and
all we really want from it is a few constants.

Also, we don't need to compile sqlite3 into the client, so let's not do that
either.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>