implement per-container idmap
Closes #632

Rename some functions to clarify what they do

Put a idmapSet into the container struct.  Currently this is by
default inherited from the daemon (meaning it is always root's
full (first) allocation, unless it is privileged).  However by
switching to idmapsets, we've now set the stage for policies and
container configs to be able to modify or add idmaps.  When that
is done we won't have to use lxc.raw to add ids any more.

Use the container idmap rather than the daemon's in the various
places we need them (rootfs shifting, etc).

Fix the id shifting at FilePut - properly shift the id into the
container rather than (a) hoping the id is mapped and (b) assuming
there is only a single map range.

When adding a range to an existing id mapping, check to make sure
that the mapping does not intersect with existing mappings.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>