Add LVM thin pool support.
Fixes #683.

Add two server configs: core.lvm_vg_name and core.lvm_thinpool_name.

If lvm_vg_name is set, use thin pool LVs for image and container
storage.

If lvm_thinpool_name is set, use that pool, otherwise create a thin pool
LV named “LXDPool" on image import and create a thin LV for that
image. Because the default setting for lvm is to not auto-extend pools,
that pool will be allocated with 100% of the remaining free space of the
volume group.

Adds test/lvm.sh with a set of tests that are run twice - once for the
default LXDPool creation and once with a pre-created test pool.
Unfortunately, the tests can not run on Travis, so they are disabled there.

On container create, creates a thin snapshot for the container storage.

This branch does not add support for snapshots or migration.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>