Update to use Go 1.3-compatible fork of go-systemd
The unsetEnv feature of the activation code in go-systemd requires
os.Unsetenv which doesn't exist until Go 1.4.

As we're currently relying on Go 1.3, I made a fork of go-systemd in
stgraber/lxd-go-systemd which has that codepath commented out (we don't
use it) and changed to print a message if someone was to use it.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>