add initial implementation of /dev/lxd
Here's an initial implementation of /dev/lxd (it doesn't do events right
now, and the userdata endpoint returns 503 since we haven't implemented it
yet in lxd proper). But, the basics are there, and this is probably enough
for people who want to start working against it to do so.

I haven't added any automated tests yet, but I did include a test program
that exits 0 if /dev/lxd works, and 1 if it doesn't:

$ lxc exec foo devlxd-client
/dev/lxd ok

Also, obviously there are some pretty brutal hacks in here to make things
work. Unfortunately, I'm not sure there's a better way to do some of this
stuff, but am open to suggestions (or better yet, patches :D).

Closes #747
Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>