exec: do not close fds until we are definately done
By closing all fds (in noninteractive case) as soon as RunCommand
is done, we occasionally miss some data still coming from the
container.  Instead, wait until stdin from the container gets an
EOF.

This stops a (on one of my systems, very easy to reproduce) failure
of the concurrency test.  To track this down I also added a simpler
test suggested by stgraber - only doing parallel execs of a single
running container.  In this way it's easy to look for any
exec-$i.out which is missing.  Tha test is also in this request.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>