move mounting of lvm vols to container start/stop
Instead of mounting a container's LV on create and unmounting it
on destroy, mount on start/stop, ensuring that we do not have
unused images mounted, and that LV-backed containers restarted
after a reboot have their mounts restored.

Adds tests for unmounting after force-stopping, but due to
busybox not working with lxc shutdown, only includes
commented-out tests for unmounting after shutdown.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>