implement device hotplug
Closes #419

It only works for veth nics, as that's all we currently support.

Fix the device type handling in the database - until now they were
being wrongly stored as strings even though the field is int.

This patch updates the list of device types, making 'none' 0
instead of 4.

The updateFromV9 comes courtesy of stgraber, as does the
documentation fix.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>