Detect container and image storage.
Fixes #827.

Use existence of the .lv symlink to detect if a container or image is
backed by an LVM LV. Use appropriate storage driver based on the storage
of the container being deleted, snapshotted or copied-from.

Also modifies delete behavior so that an error from the storage driver
does not stop the container from being deleted in the db.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>