bump the fdleak test to >5 fds
As I suspected, Gosched() call doesn't really help us: it just explicitly
queues the finalizers, but doesn't actually block until they're finished.
So, it makes the race we're seeing in the fdleak test much less likely, but
it still happens on occasion.

Instead, let's do the test in a loop five times, and check that we haven't
leaked >5 fds.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>