Add lvm native snapshot support
Remove storage start/stop around container restore. It was there to
ensure LVM containers were mounted to do an rsync, and is unnecessary
when using lvm snapshots, which don't require mounting anything.

Due to the small character set allowed for LVM LV names, we map the
snapshot delimiter '/' to '-'. Then, to allow a container named
'cream-cheese' to coexist with a snapshot named 'cheese' of a
container named 'cream' (aka 'cream/cheese'), we map single '-' to
'--' and '/' to '-'. So the container 'cream-cheese' has an LV
'cream--cheese'. Don't look at the devicemapper device name for that
one.

We rename existing lv-backed snapshots in a 'fake' db schema
upgrade to the new scheme.

Because LVM does block-level snapshots, we need to stop and unmount
the LV before taking a snapshot in order to have consistent state.

This branch only does this for LVM containers and doesn't provide a way
to warn users.

Signed-off-by: Michael McCracken <michael.mccracken@gmail.com>