move raw.lxc config to applyPostDeviceConfig()
As per bug #588 and fix #635, raw.lxc settings have to be applied after
the network devices config are set.

However, we also want to ensure that profile config is overridden by
container-specific config, which requires a second call to applyConfig
with the baseConfig after the profiles are applied. This however can't
be at the end, so we split the application of raw.lxc settings into a
separate function. This eliminates the need to track in applyConfig
whether or not the config came from a profile.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>