remove restriction on renaming a running container
This is fine from a LXD point of view. The spec promises that we'll change the
hostname, but we currently don't do that for non-running containers, so there's
no reason to prevent this with running containers.

We might consider removing the hostname promise all together, since it seems a
little invasive.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>