Makefile: run go get twice
This is needed to workaround what looks like a bug in the dependency
resolver when also building tests.

Without this, in some cases, go get will fail after downloading most
dependencies, saying that it can't find another one of the dependencies.
A subsequent retry grabs the missing bits and go get completes
succesfuly.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>