Implement shutdown command (and some cleanup)
This implements the shutdown command as a lxd argument.

When called, SIGPWR is sent to LXD by fetching the daemon's PID over the
API, we then wait for LXD to stop responding before returning.

This is meant to be used by init systems to perform a clean shutdown of
LXD on host shutdown.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>