Do not freeze containers when taking LVM Snapshots
LVM takes care of this for us - it sends a device mapper 'suspend'
message to the origin thin volume before taking the (block-level)
snapshot. This ends up calling freeze_super and freeze_fs, which
quiesces the filesystem and allows for a consistent snapshot.

freeze_fs is supported on ext3/4 and btrfs, and others.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>