get rid of migration subpackage
We'll need this to fix migration, since we need to call the apparmor/seccomp
primitives, and those live in lxd/, and we can't do circular imports. (Moving
the security stuff to a security/ package is also nontrivial, since that
depends on containerLXD, which itself should be in a separate package, etc.
etc.)

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>