delay check for subuids
if we are creating a privileged container, we shouldn't fail
creation because of there being no subuids.  But we can't
easily know at the start of containers_post whether the container
is privileged.  So delay the check until we have the info, and
until we definately need it.

As a downside, an error message won't be seen until we've
downloaded the image.  To change that we'd have to do c.init()
before checking for and downloading the image.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>