tests: Don't hardcode ports
This is the first step of making it possible to run the testsuite
multiple times in parallel.

There is a potential race between the moment we kill the netcat binding
our future port and LXD binding it itself but since I'm using random
ports rather than iterating sequentially, clashes should be extremely
unlikely.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>