Move mountShared from container init to container start
This sets up the shmounts path for the container, since we never write
to it for a stopped container, there's no reason for it to be available
before start. Also this was causing the container's lxc config to be
appended to every time a container would be accessed.

This also has the side effect to fix running the go tests as an
unprivileged user.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>