Make it possible to set/unset the public flag
Done through image edit.

This also fixes the images API to match the spec, unfortunately this
means breaking some backward compatibility. An older client will not be
able to interact with a newer server.

The new client will however still be able to interact with an older
server, at least for the time being.

Closes #976

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>