Fix network device hotplug
This makes network device hotplug respect hwaddr, generate new MACs when
needed and store to volatile storage when needed.

It also changes the host prefix to veth so Network Manager doesn't freak out.

Note that the raw SQL is unfortunately necessary due to all of this
running in the middle of a ConfigReplace transaction...

Closes lxc/nova-compute-lxd#46

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>