tests: Don't fail on kernel race conditions
On some systems it seems to be possible to run into a case where the MAC
address of a device isn't actually set for a while.

This is likely a kernel issue as a veth pair device should always have a
MAC, but it's shown up a few times, especially in the Ubuntu CI lab and
is annoying, so let's ignore that case since it's clearly not a LXD bug.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>