Implement unix char and block devices.
Closes #1200
Closes #483

If we have the ability to mknod, then we create a device in the
containerdir and bind mount that into the container.  This allows
us to uidshift the device to the container root.  If we can not
mknod, then we try to bind mount the provided device pathname
from the host into the container.  This supports nested containers.

Create a helper for inserting a mount into a container, since
we need it both for adding disk and unix block/char devices.

Add a testcase.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>