Implement remote publish
This is a first pass on remote publish, things could be optimized a bit
by having the source server export the tarball immediately, skipping the
creation of a temporary DB record and the potential creation of a
zfs/btrfs/lvm volume.

Such an optimization should happen as a websocket or plain http link
against the operation.

Closes #796

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>