make blockdev mounts work
Closes #906

A block device /dev/loop0 is mounted onto $containerdir/blk.-dev-loop0.
From there it is mounted into the container's shmounts directory,
and then move-mounted inside the container.

TODO
We will want to add a post-stop hook to all lxd containers which
unmounts any "$containerdir/blk.*" mounts.  This is done automatically
when you lxc stop a container, but if the container does a 'shutdown'
it is not.  It'll still happen at the next start so this isn't
urgent, but not doing so can on weird systemd eventually leave a lot
of mounts around.  I didn't do it here as I wasn't quite sure how
best to ship that from upstrema lxd.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>