migration: migrate snapshots too
This commit also moves snapshots when copying containers. Note that it is
pretty dumb: it also copies snapshots in the hot path of migraiton, which
is a bad deal. Further, it will only copy things between directory storage
backends; this adds the framework necessary for fs specific backend
send/receive, to be used in later patches.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>