Add events interface
This adds the events interface and eventSend function that can be hooked
wherever we want to send events back to the client.

Also adds a basic "lxc monitor" command (hidden) to show messages as
they arrive.

Note that some operation events aren't currently handled as the
operation subsystem doesn't provide a single codepath for operation
completion and cancelation, making it hard to send an event for those.

Also, this doesn't hook in our logging system yet.

Closes #167

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>