Keep operations around for 5 seconds
This fixes any potential race where a client or even a goroutine are
attempting to get hold of the completed operation and fail because it
just got deleted and garbage collected.

A longer delay would be nicer as that'd allow to see what happened over
the past 30s or minute without causing too much memory use, but doing so
would require adding a matching sleep time in the fdleak test, so stick
to 5 seconds for now.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>