drop migratable profile
As of kernel 4.4/criu 1.8, users will be able to checkpoint/restore
containers which are in user namespaces and using seccomp filters (as well
as STRICT mode). Further, we already set tty = 0 (as required by liblxc)
and we don't use the console, so let's just hardcode these things all
together and ditch the profile that's required.

That is, every LXD container (with recent enough kernel/criu) is migratable
(potentially, modulo whatever CRIU currently supports) by default.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>