Replace goproxy by minimal http proxy
This gets us rid of goproxy and all its dependencies.

The result is equivalent for HTTP proxying, we do loose support for
proxying of other protocols but since lxd-bridg-proxy has never been
turned on for anyone, that's fine for now.

We'll set it as an HTTP proxy by default and if there's strong demand,
add HTTPs proxying then (CONNECT support).

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>