Implement basic CPU scheduler and limits.cpu
This adds a goroutine that monitor uevents for CPU hotplug events as
well as hook into the container lifecycle.

On CPU add/remove or container start/stop/update, a basic scheduler is
invoked which updates the cpuset pinning of all containers, keeping the
fixed allocations on their requested CPU and spreading the rest across
all CPUs.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>