rsync: wait for websockets to be done
Since we're going to re-use the socket, now we need to wait for the first
send/receive to be entirely complete before we move on to the next one.

It's a little bit ugly that we have to wait for both readDone and writeDone
everywhere, we could make a /third/ goroutine in WebsocketMirror to wait on
both readDone and writeDone and send a message over a third channel when
both are complete. That would make the user API nicer, but it is a bit more
convoluted, and I'm happy to send a patch that does that if it seems
better, though.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>