tests: In ephemeral test, ignore failure to reboot
Some versions of busybox seem to sometimes ignore the reboot command,
leading to random test failures.

What we're testing here isn't busybox, it's whether lxc regresses as far
as ephemeral support and starts removing containers it shouldn't.

So just ignore the restart failure as we used to before I "fixed" the
test a couple weeks ago.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>