copy: add logging for intermediate errors in --debug
The problem here is that we don't have a good way to distinguish between
the "the network didn't work at this address" error and the "the operation
actually failed" error.

The end result is that if the first address was the right one, but the
operation failed, we try to reconnect at an invalid address and end up with
a "websocket: bad handshake" error, with no way to debug what the actual
problem was. At least this way, we can have people pass --debug and see
what the real error was if they're hitting this case.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>