migration: don't hardcode the order of snapshot/container xfer
In the rsync case we send the container first (although arguably we could
get some benefit by not syncing snapshots independently, so this may change
in the future). In the zfs case, we send the container last. Let's not
enforce this, and instead use an IsSnapshot() method to detect wheter or
not the thing we're sending is a snapshot or the actual contianer, so
things can be sent in any order.

Also, this adds a small special case when transferring a snapshot as a
container.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>