zfs: add an fs-specific transfer mechanism for zfs
This commit implements a filesystem specific transfer mechanism for zfs.
The one known caveat with this is that it won't necessarily work with live
migrations since we don't do a delta send after the live snapshot is taken.
I'm still trying to figure out what the best way to do that is, given all
the restrictions with zfs.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>