lvm: Harden all LVM and mount calls
It looks like on busy systems LVM will ocassionaly think a volume is in
use when it shouldn't.

There are no obvious ways to confirm whether it's in fact the case, but
at least in our testsuite, it's not. As a result, attempt every LVM and
mount/umount actions up to 10 times in 500ms increment.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>