zfs: hopefully fix some more races
Let's try to unmount the zfs volume in two cases: 1. when it is a
mountpoint (so we know for sure it is mounted), or 2. when zfs thinks it is
mounted (via 'zfs mount'). Presumably zfs will not refuse to recv something
if `zfs mount` says it's unmounted, but who really knows.

i.e. I'm not sure if this will actually fix the race, but I hope so :)

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>