travis: use "system" version of go vet
When we pull from trunk, if the repo for go vet makes a change that only
compiles in e.g. 1.5, then all the tests for 1.4 and 1.3 fail, since we try
to pull the source version of go vet for each go version. Instead, let's
use the "system" version (i.e. the version of go vet distributed with that
toolchain), so we don't have build issues.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>