Always set a root disk entry
In preparation for container disk quotas, we must ensure that all
containers have a disk device entry for /.

This commit does the following:
 - Introduce a new type of config/devices checks for expanded configuration
 - Prevent removing a disk entry for path /
 - Create a blank disk entry for path / at container creation time
   (if none is inherited from a profile).
 - Update all existing containers to have a valid disk entry

Note that while this code adds the disk entry directly to the container,
it's perfectly possible for an administrator to choose to have only a
single one shared with all containers through profile. In such case,
that disk entry must be added to a profile and then the individual disk
entries from the container can be removed.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>