tests: Improve performance of concurrent test
Instead of a fixed 50 concurrent containers which will on some system
cause such a load that we start reaching our timeouts, lets go with a
more realistic 8x the number of available CPU cores (and so number of go
threads).

Also to speed things up, use lxc info instead of lxc list to query
container status.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>