Fix race condition in event interface
When flooded by events on a SMP system, it was technically possible to
have two threads write to the same event listener, leading to corrupted
websocket frames or json content.

Introduce a per-listener lock that's acquired during the time a message
is pushed over the wire and update the client code to better report
errors.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>