Initial work on a LXC to LXD script
This new script supports migrating most privileged containers to LXD
containers. It supports all the networking and bind-mount options that
LXD itself does, it also properly sets the nesting mode, boot options
and environment variables.

Various options are available to migrate all or just some specific
containers, specify custom lxcpath and lxdpath, make a copy of the
container or move the data and delete the source, ...

Security knobs which LXD doesn't provide cause an error so containers do
not accidently get migrated to a configuration which is less secure.

Unprivileged containers will require some extra work to unmap and remap
them to the LXD range.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>