move usage() message before error
My eyes are still at the prompt, so I have to scroll back visually and
parse all of this output to find the error, which is actually the first
line printed. Instead, let's just print it as the last thing, which means
that it's probably the first thing I see.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>