Rework container state API
Returning all the runtime state all the time is very costly, it's
actually better to only query and return it on demand.

So only return the basic container state (running/stopped/...) with the
container info and require accessing /state for anything else.

This will be even more important as we add cgroup resource reporting and
other bits of data in there soon.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>