Refactor the GenCert function so it can be reused.
Juju was working around the fact that GenCert always wrote directly to disk. Instead split the
function into 2 functions. One that just generates the certificate bytes, and another one
that will write that content into files.
Add a test for the helper function. It is unfortunately a bit slow (2s).
Also, suppress printing all the lines that were read by ReadLastNLines, as it makes it
hard to see what the actual test suite is doing.

Signed-off-by: John Arbash Meinel <john@arbash-meinel.com>